// frontend/src/app/checkout/types.ts
export interface AddressFormData {
    fullName: string;
    phone: string;
    email?: string;
    province: string;
    district: string;
    ward: string;
    address: string;
    notes?: string;
  }
  
  export interface OrdererInfo {
    fullName: string;
    phone: string;
    email?: string;
  }
  